Train 52 - deletePets

+++ src/service/petService.ts
@@ -56,6 +56,10 @@ export const deletePets = async (
   id?: string[],
   petTag?: string[]
 ): Promise<void> => {
+  if (!id && !petTag) {
+    logger.warn('Nither id nor petTag has been provided');
+    throw new Error('Either id or petTag must be provided');
+  }
   if (id) {
     logger.info(`Deleting pets by IDs`);
     await Pets.deleteMany({
